+++ sdk/model/types_test.go
@@ -38,7 +38,7 @@ func TestNewStreamOptions_Defaults(t *testing.T) {
 	opts := NewStreamOptions()
 	assert.Empty(t, opts.Model)
 	assert.Equal(t, ThinkingOff, opts.ThinkingLevel)
-	assert.Equal(t, int64(0), opts.MaxTokens)
+	assert.Equal(t, 0, opts.MaxTokens)
 }
 
 func TestNewStreamOptions_FunctionalOptions(t *testing.T) {
@@ -49,12 +49,12 @@ func TestNewStreamOptions_FunctionalOptions(t *testing.T) {
 	)
 	assert.Equal(t, "claude-opus-4-7", opts.Model)
 	assert.Equal(t, ThinkingHigh, opts.ThinkingLevel)
-	assert.Equal(t, int64(8192), opts.MaxTokens)
+	assert.Equal(t, 8192, opts.MaxTokens)
 }
 
 func TestNewStreamOptions_PartialOptions(t *testing.T) {
 	opts := NewStreamOptions(WithThinkingLevel(ThinkingMedium))
 	assert.Empty(t, opts.Model)
 	assert.Equal(t, ThinkingMedium, opts.ThinkingLevel)
-	assert.Equal(t, int64(0), opts.MaxTokens)
+	assert.Equal(t, 0, opts.MaxTokens)
 }
